html,css,js 수정 register,update,delete 기능 수정

+++ src/main/resources/static/js/app/main.js
@@ -11,51 +11,32 @@ var index ={
         });
         $('#btn_delete').on('click',function ()
         {
-            _this.delete();
+            if(confirm("정말로 삭제를 진행하시겟습니까?")==true)
+            {
+                _this.delete();
+            }
+
         });
         $('#btn_search').on('click',function()
         {
             _this.search();
         });
+        $('#btn_register').on('click',function(){
+            _this.register();
+        });
+        $('#check_employees').on('click',function(){
+            if($("input:checkbox[id='check_employees']").prop("checked"))
+            {
+                $("input[type=checkbox]").prop("checked",true);
+            }else
+            {
+                $("input[type=checkbox]").prop("checked",false);
+            }
+        })
     },
     save : function()
     {
-        var data = {
-            orgid:$('#orgid').val(),
-            companyid:$('#companyid').val(),
-            employeenumber:$('#employeenumber').val(),
-            inspectortype:$('#inspectortype').val(),
-            krname:$('#krname').val(),
-            departmentcode:$('#departmentcode').val(),
-            positioncode:$('#positioncode').val(),
-            upperemployeenumber:$('#upperemployeenumber').val(),
-            leaderyn:$('#leaderyn').val(),
-            effectivestartdate:$('#effectivestartdate').val(),
-            effectiveenddate:$('#effectiveenddate').val(),
-            email:$('#email').val(),
-            phonenumber:$('#phonenumber').val(),
-            useyn:$('#useyn').val(),
-            remarks:$('#remarks').val(),
-            createdby:$('#createdby').val(),
-            creationdate:$('#creationdate').val(),
-            lastupdatedby:$('#lastupdatedby').val(),
-            lastupdatedate:$('#lastupdatedate').val()
-        };
-        $.ajax({
-            type:'POST',
-            url:'/api/register',
-            dataType:'json',
-            contentType:'application/json;charset=utf-8',
-            data:JSON.stringify(data)
-        }).done(function(msg)
-        {
-            alert(msg.responseText);
-            window.location.href='/';
-        }).fail(function (error){
-            alert(error.responseText);
-            window.location.href='/';
-        })
-
+        window.location.href='/employee/register';
     },
     update : function()
     {
@@ -98,21 +79,46 @@ var index ={
     },
     delete : function()
     {
-        var employeenumber = $('#employeenumber').val();
-
+        var check_cnt = document.getElementsByName("check_employee").length;
+        var employeenumbers = new Array();
+        for(var i =0;i<check_cnt;i++)
+        {
+            if (document.getElementsByName("check_employee")[i].checked == true) {
+                // alert(document.getElementsByName("check_employee")[i].value);
+                employeenumbers.push({"employeenumber":document.getElementsByName("check_employee")[i].value});
+            }
+        }
+        var jsonString = JSON.stringify(employeenumbers)
         $.ajax({
             type:'DELETE',
-            url:'/api/delete/'+employeenumber,
+            url:'/api/delete',
             dataType:'json',
-            contentType:'application/json; charset=utf-8'
+            contentType:'application/json; charset=utf-8',
+            data:jsonString
         }).done(function()
         {
-            alert('사원이 삭제되었습니다.');
+            alert("삭제 완료되었습니다.");
             window.location.href='/';
-        }).fail(function (error)
+        }).fail(function(error)
         {
-            alert(JSON.stringify(error));
+            alert(error.responseText);
+            window.location.href='/';
         })
+        // var employeenumber = $('#employeenumber').val();
+
+        // $.ajax({
+        //     type:'DELETE',
+        //     url:'/api/delete/'+employeenumber,
+        //     dataType:'json',
+        //     contentType:'application/json; charset=utf-8'
+        // }).done(function()
+        // {
+        //     alert('사원이 삭제되었습니다.');
+        //     window.location.href='/';
+        // }).fail(function (error)
+        // {
+        //     alert(JSON.stringify(error));
+        // })
     },
     search : function()
     {
@@ -120,13 +126,44 @@ var index ={
         var keyword = $('#seacrchKeyword').val();
 
         window.location.href='/employee/search?title='+title+'&keyword='+keyword;
+    },
+    register : function()
+    {
+        var data = {
+            orgid:$('#orgid').val(),
+            companyid:$('#companyid').val(),
+            employeenumber:$('#employeenumber').val(),
+            inspectortype:$('#inspectortype').val(),
+            krname:$('#krname').val(),
+            departmentcode:$('#departmentcode').val(),
+            positioncode:$('#positioncode').val(),
+            upperemployeenumber:$('#upperemployeenumber').val(),
+            leaderyn:$('#leaderyn').val(),
+            effectivestartdate:$('#effectivestartdate').val(),
+            effectiveenddate:$('#effectiveenddate').val(),
+            email:$('#email').val(),
+            phonenumber:$('#phonenumber').val(),
+            useyn:$('#useyn').val(),
+            remarks:$('#remarks').val(),
+            createdby:$('#createdby').val(),
+            creationdate:$('#creationdate').val(),
+            lastupdatedby:$('#lastupdatedby').val(),
+            lastupdatedate:$('#lastupdatedate').val()
+        };
+        $.ajax({
+            type:'POST',
+            url:'/api/register',
+            dataType:'json',
+            contentType:'application/json;charset=utf-8',
+            data:JSON.stringify(data)
+        }).done(function(msg)
+        {
+            alert(msg.responseText);
+            window.location.href='/';
+        }).fail(function (error){
+            alert(error.responseText);
+            window.location.href='/';
+        })
     }
 };
 index.init();
-// function checkAll()
-// {
-//     if($('employee_check').is('check'))
-//     {
-//
-//     }
-// }
